Extract `NumberVariance` to standalone class
