+++ src/main/kotlin/com/github/oberdiah/deepcomplexity/staticAnalysis/BundleExtensions.kt
@@ -6,8 +6,8 @@ fun <T : Number> Bundle<T>.into(): NumberSet<T> =
 fun Bundle<Boolean>.into(): BooleanSet =
     this as BooleanSet
 
-fun <T : Number> VarianceBundle<T>.into(): NumberSet.NumberVariance<T> =
-    this as NumberSet.NumberVariance<T>
+fun <T : Number> VarianceBundle<T>.into(): NumberVariance<T> =
+    this as NumberVariance<T>
 
 fun VarianceBundle<Boolean>.into(): BooleanSet.BooleanVariance =
     this as BooleanSet.BooleanVariance
