+++ src/main/kotlin/com/github/oberdiah/deepcomplexity/staticAnalysis/NumberSet.kt
@@ -27,42 +27,6 @@ class NumberSet<T : Number>(
     override fun getIndicator(): SetIndicator<T> = ind
     val clazz: KClass<T> = ind.clazz
 
-    class NumberVariance<T : Number>(private val set: NumberSet<T>) : VarianceBundle<T> {
-        override fun getIndicator(): SetIndicator<T> = set.getIndicator()
-        override fun <Q : Any> cast(newInd: SetIndicator<Q>): VarianceBundle<Q>? {
-            if (newInd !is NumberSetIndicator<*>) {
-                return null
-            }
-
-            @Suppress("UNCHECKED_CAST")
-            return set.cast(newInd)?.let { NumberVariance(it.into()) } as NumberVariance<Q>?
-        }
-
-        fun negate(): NumberVariance<T> = set.negate().withEphemeralVariance().into()
-
-        override fun union(other: VarianceBundle<T>): VarianceBundle<T> =
-            NumberVariance(set.union(other.into().set) as NumberSet<T>)
-
-        override fun collapse(): Bundle<T> = set
-
-        fun isOne(): Boolean = set.isOne()
-        fun arithmeticOperation(
-            other: NumberVariance<T>,
-            operation: BinaryNumberOp
-        ): NumberVariance<T> = set.arithmeticOperation(other.set.into(), operation).withEphemeralVariance().into()
-
-        fun comparisonOperation(
-            other: NumberVariance<T>,
-            operation: ComparisonOp
-        ): BooleanSet.BooleanVariance = set.comparisonOperation(other.set, operation).withEphemeralVariance().into()
-
-        fun evaluateLoopingRange(
-            changeTerms: ConstraintSolver.EvaluatedCollectedTerms<T>,
-            valid: NumberSet<T>
-        ): NumberVariance<T> = set.evaluateLoopingRange(changeTerms, valid).withEphemeralVariance().into()
-
-    }
-
     override fun toString(): String = ranges.joinToString {
         it.toString()
     }
